+++ core/CHANGELOG.md
@@ -9,6 +9,8 @@
 * Refactor the new intializer's `main` method, which had become 550 lines long, into separate classes
 * Run tests in parallel
 * Clean up some bugs in DockerManager where `context.Background` was getting used where it shouldn't
+* Added test to make sure the IP placeholder string replacement happens as expected
+* Actually mount the test volume at the location the user requests in the `AddService` Kurtosis API endpoint
 
 # 0.9.0
 * Change ConfigurationID to be a string